Simple Tip Calculator
Another simple project i have built to ensure i am implementing the skills i am learning such as the "children" prop which i have come to learn of most recently, i made an effort to include them in this even though they are not necessary for such a small application i wanted to make use of them to really grasp the use cases. as there are more features such as moving up state and derived state which i feel i have really gotten to grips with in this project. Its by no means a portfolio project but i think it is cleanly written and makes use of several core features and benefits of react.